# docs/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Idenplane - Open Source Identity and Access Management
  description: |
    Idenplane is an open-source authentication and authorization server implementing
    the OAuth 2.0 and OpenID Connect protocols.

    ## Authentication

    This API supports two authentication mechanisms:
    1. **Admin API Key**: Include `x-admin-api-key: YOUR_KEY` header
    2. **Bearer Token**: Include `Authorization: Bearer TOKEN` header (for user-facing endpoints)

    ## Versioning

    API endpoints are versioned using URI versioning (`/api/v1/` prefix).

    ## Rate Limiting

    Some endpoints are rate-limited. See specific endpoint documentation for details.
  version: 1.0.0
  contact:
    name: Idenplane Project
    url: https://idenplane.dev
    email: support@idenplane.dev
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
security:
  - admin-api-key: []
  - bearerAuth: []
...
paths:
...
  /realms/{realmName}/protocol/openid-connect/auth:
    get:
      tags:
        - OAuth
      summary: Authorization endpoint (code flow)
      description: |
        OAuth 2.0 Authorization Endpoint. Initiates the authentication flow.
        If the user has an active session, they may be redirected directly to the redirect_uri.
        Supports PKCE (code_challenge and code_challenge_method parameters).
      operationId: authorize
      parameters:
        - name: realmName
          in: path
          required: true
          schema:
            type: string
        - name: client_id
          in: query
          required: true
          schema:
            type: string
          description: OAuth client ID
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
          description: OAuth redirect URI
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum:
              - code
          description: OAuth response type
        - name: scope
          in: query
          schema:
            type: string
            default: openid
          description: OAuth scopes (space-separated, must include 'openid')
        - name: state
          in: query
          schema:
            type: string
          description: CSRF state parameter
        - name: nonce
          in: query
          schema:
            type: string
          description: Nonce for ID token
        - name: code_challenge
          in: query
          schema:
            type: string
          description: PKCE code challenge
        - name: code_challenge_method
          in: query
          schema:
            type: string
            enum:
              - S256
              - plain
          description: PKCE code challenge method
        - name: prompt
          in: query
          schema:
            type: string
            enum:
              - none
              - login
              - consent
              - select_account
          description: OIDC prompt parameter
        - name: acr_values
          in: query
          schema:
            type: string
          description: Requested Authentication Context Class Reference values
        - name: login_hint
          in: query
          schema:
            type: string
          description: Pre-fill the username field
      responses:
        '302':
          description: Redirect to login page or redirect_uri with authorization code
        '400':
          description: Invalid request - missing or invalid OAuth parameters
...
  /realms/{realmName}/protocol/openid-connect/revoke:
    post:
      tags:
        - Tokens
      summary: Token revocation (RFC 7009)
      description: Revokes an access or refresh token.
      operationId: revokeToken
      parameters:
        - name: realmName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  description: The token to revoke
                token_type_hint:
                  type: string
                  enum:
                    - access_token
                    - refresh_token
                  description: Hint about the type of the token
                client_id:
                  type: string
                  description: OAuth client ID
                client_secret:
                  type: string
                  description: OAuth client secret
      responses:
        '200':
          description: Token successfully revoked (or was already invalid)
        '400':
          description: Bad request - missing token parameter
        '401':
          description: invalid_client - client authentication failed
...
components:
  securitySchemes:
    admin-api-key:
      type: apiKey
      name: x-admin-api-key
      in: header
      description: Admin API key for administrative endpoints

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for user-facing endpoints
...